requirements.txt
# Lenovo Multi-Agent Sales Intelligence System
# Python 3.10+

# Core orchestration & LLM
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-openai>=0.2.0

# Semantic search (optional but recommended)
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4

# Frontend
streamlit>=1.28.0

# Utilities (often pulled in by above)
numpy>=1.24.0
